Refactor: entity.mapping 패키지에 속한 PlaceCategory를 entity패키지로 변경 (#175)
