+++ src/main/java/com/ilta/solepli/domain/solmap/service/SolmapService.java
@@ -22,10 +22,10 @@ import lombok.extern.slf4j.Slf4j;
 import com.ilta.solepli.domain.category.entity.QCategory;
 import com.ilta.solepli.domain.category.repository.CategoryRepository;
 import com.ilta.solepli.domain.place.entity.Place;
+import com.ilta.solepli.domain.place.entity.PlaceCategory;
 import com.ilta.solepli.domain.place.entity.QPlace;
+import com.ilta.solepli.domain.place.entity.QPlaceCategory;
 import com.ilta.solepli.domain.place.entity.QPlaceHour;
-import com.ilta.solepli.domain.place.entity.mapping.PlaceCategory;
-import com.ilta.solepli.domain.place.entity.mapping.QPlaceCategory;
 import com.ilta.solepli.domain.place.repository.PlaceRepository;
 import com.ilta.solepli.domain.review.dto.response.ReviewDetail;
 import com.ilta.solepli.domain.review.entity.QReview;
